pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Satoaki Ishihara"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.35.10"
load-dotenv = "^0.1.0"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
langchain-openai = "^0.1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
